backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.0
pydub==0.25.1
python-dotenv==1.0.1
redis==5.0.7
celery==5.4.0

